+++ src/rules/__tests__/contains.test.ts
@@ -1,4 +1,4 @@
-import { checkClue } from "../../rules";
+import checkClue from "../../check-clue";
 
 test("contains", () => {
   expect(checkClue("A screenshot of a dog.")).toMatchInlineSnapshot(`
