+++ src/rules/__tests__/end-with.test.ts
@@ -1,4 +1,4 @@
-import { checkClue } from "../../rules";
+import checkClue from "../../check-clue";
 
 test("endWith", () => {
   expect(checkClue("A screenshot of a dog.jpg")).toMatchInlineSnapshot(`
