+++ src/rules/char-length.ts
@@ -1,5 +1,5 @@
-import { Rule } from "../clues";
-import { createSuggestion } from "../rules";
+import { Rule } from "./index";
+import { createSuggestion } from "../check-clue";
 import { Alt } from "../index.js";
 
 export function checkLength(alt: Alt) {
