+++ src/rules/end-with.ts
@@ -1,4 +1,4 @@
-import { Rule } from "../clues";
+import { Rule } from "./index";
 
 export default function atlEndsWith(): Rule {
   return {
