+++ src/rules/exact-match.ts
@@ -1,4 +1,4 @@
-import { Rule } from "../clues";
+import { Rule } from "./index";
 
 export default function exactMatch(): Rule {
   return {
