+++ src/rules/image-link.ts
@@ -1,4 +1,4 @@
-import { Rule } from "../clues";
+import { Rule } from "./index";
 
 export default function imageLink(): Rule {
   return {
