+++ src/rules/no-alt.ts
@@ -1,4 +1,4 @@
-import { Rule } from "../clues";
+import { Rule } from "./index";
 
 export default function noAlt(): Rule {
   return {
